requirements.txt
# Web framework
flask==2.3.3
flask-cors==4.0.0

# Web scraping
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.11.2
webdriver-manager==4.0.0

# Text processing and NLP
nltk==3.8.1
spacy==3.6.1
sentence-transformers==2.2.2

# Vector database
chromadb==0.4.15
langchain==0.0.305
langchain-community==0.0.10

# Utilities
python-dotenv==1.0.0
tqdm==4.65.0
pandas==2.1.0
numpy==1.25.2
html2text==2024.2.26

# Development
pytest==7.4.2
black==23.7.0
flake8==6.1.0